Pick a different name
Aether CSS is probably a bit too close to https://ether.thescenery.co/ which is
in the same space.

Substrate CSS doesn't seem encumbered by other projects, and it's delightfully
close to Substance Lab, and I like the connotation that it's something extracted
from other projects (which it is).

Alas, it is longer to write `@include substrate-header`, though.